+++ backend/src/dtos/clients.dto.ts
@@ -1,5 +1,9 @@
 import { ApiKeyStatus } from "../utils/enums/api-key-status.enum";
 
+export interface ClientsCreateDTO {
+    name: string
+}
+
 export interface ClientsStatusUpdateDTO {
     status: ApiKeyStatus
 }