+++ backend/src/repositories/clients.repository.ts
@@ -14,7 +14,7 @@ class ClientsRepository {
         this.table = 'clients';
     }
 
-    async findByKey(hash: string): Promise<Clients | IRepoError | null> {
+    async findByActiveKey(hash: string): Promise<Clients | IRepoError | null> {
         const sql = `SELECT * FROM ${this.table} WHERE api_key_hash = $1 AND status = $2;`;
         const values = [hash, ApiKeyStatus.ACTIVE];
         const db = DBConnection.getInstance();
@@ -27,11 +27,11 @@ class ClientsRepository {
         } catch(err: any) {
             // TODO(yqni13): logging
             if(secrets.ENV_MODE === EnvMode.DEV || secrets.ENV_MODE === EnvMode.TEST) {
-                console.log("DB ERROR ON SELECT (Clients Repository, findByKey): ", err);
+                console.log("DB ERROR ON SELECT (Clients Repository, findByActiveKey): ", err);
             }
             await db.close(client);
             return {
-                method: 'support_clients_findByKey',
+                method: 'support_clients_findByActiveKey',
                 error: err
             }
         }
@@ -60,15 +60,14 @@ class ClientsRepository {
         }
     }
 
-    async create(id: string, name: string, hash: string): Promise<Clients | IRepoError> {
+    async create(id: string, hash: string, dto: Partial<Clients>): Promise<Clients | IRepoError> {
         const timeStamp = Utils.getTimestampWithOffsetInfo(new Date());
         const sql = `INSERT INTO ${this.table}
         (client_id, name, api_key_hash, status, last_use, last_modified, created_on)
         VALUES ($1, $2, $3, $4, $5, $6, $7)
         RETURNING *;
         `;
-        const values = [id, name, hash, ApiKeyStatus.ACTIVE, timeStamp, timeStamp, timeStamp];
-        
+        const values = [id, dto.name, hash, ApiKeyStatus.ACTIVE, timeStamp, timeStamp, timeStamp];
         const db = DBConnection.getInstance();
         let client: any;
         try {