+++ backend/src/services/clients.service.ts
@@ -1,18 +1,24 @@
-import { ClientsCreateResponseDTO, ClientsExistResponseDTO, ClientsLastUseResponseDTO, ClientsStatusResponseDTO, ClientsStatusUpdateDTO } from "../dtos/clients.dto";
+import {
+    ClientsCreateDTO,
+    ClientsCreateResponseDTO,
+    ClientsExistResponseDTO,
+    ClientsLastUseResponseDTO,
+    ClientsStatusResponseDTO,
+    ClientsStatusUpdateDTO
+} from "../dtos/clients.dto";
 import { IRepoError } from "../repositories/interfaces/error.repository.interface";
 import clientsModel from '../models/clients.model';
 import * as Utils from '../utils/common.utils';
 import clientsRepository from '../repositories/clients.repository';
 import { Clients } from "../repositories/interfaces/clients.entity.interface";
 
-
 class ClientsService {
     /**
      * @description Usage for apikey authentication in auth.middleware.ts.
      */
-    async findByKey(key: string): Promise<ClientsExistResponseDTO | IRepoError | null> {
+    async findByActiveKey(key: string): Promise<ClientsExistResponseDTO | IRepoError | null> {
         const hash = Utils.mapKeyToHash(key);
-        let result = await clientsRepository.findByKey(hash);
+        let result = await clientsRepository.findByActiveKey(hash);
         result = !result || Utils.isIRepoError(result)
             ? null
             : (clientsModel.mapObjToApi(result as ClientsExistResponseDTO)) as ClientsExistResponseDTO;
@@ -30,10 +36,10 @@ class ClientsService {
         }
     }
 
-    async createClient(name: string): Promise<ClientsCreateResponseDTO | IRepoError | null> {
+    async createClient(dto: ClientsCreateDTO): Promise<ClientsCreateResponseDTO | IRepoError | null> {
         const id = Utils.generateUUID();
         const apiKey = clientsModel.generateApiKeyObj();
-        let result = await clientsRepository.create(id, name, apiKey.keyHash);
+        let result = await clientsRepository.create(id, apiKey.keyHash, dto);
         if(!result) {
             return result;
         } else if(Utils.isIRepoError(result)) {