+++ backend/src/utils/errorCodes.utils.ts
@@ -18,5 +18,6 @@ export const ErrorCodes: any = {
     DBEmptyException: 16,
     MaintenanceException: 17,
     MalformedApiKeyException: 18,
-    ApiKeyMissingException: 19
+    MissingApiKeyException: 19,
+    InvalidApiKeyException: 20
 };